git-annex: fix the build in LTS 12.x
git-annex depends on 'esqueleto', which is not actively maintained. As a
result, 'esqueleto' fails to compile with recent versions of 'persistent',
which in turns requires us to link with old versions of 'conduit' and so on and
so forth. It's a disaster. I wasted several hours of my life trying to make
that build work, but eventually I gave up and dropped support for S3 and for
the Web app from git-annex to get rid of the most problematic dependencies.
That reduced version of git-annex is still difficult enough to compile, but at
least it's possible.